# openapi/schema/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "MonetaryAmountSchema" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UseNativeTypeType" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
      "UseSchemaImplementationImpl" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          }
        }
      },
      "UseSchemaImplementationType" : {
        "$ref" : "#/components/schemas/UseSchemaImplementationImpl",
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          },
          "internalValue" : {
            "type" : "string"
          },
          "composite" : {
            "type" : "boolean"
          },
          "null" : {
            "type" : "boolean"
          }
        }
      },
      "UseSchemaRefType" : {
        "$ref" : "#/components/schemas/MonetaryAmountSchema",
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          },
          "internalValue" : {
            "type" : "string"
          },
          "composite" : {
            "type" : "boolean"
          },
          "null" : {
            "type" : "boolean"
          }
        }
      }
    }
  },
  "paths" : {
    "/schema/UseNativeTypeType" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UseNativeTypeType"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          }
        },
        "summary" : "Third",
        "tags" : [ "Greeting Resource" ]
      }
    },
    "/schema/UseSchemaImplementationType" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UseSchemaImplementationType"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "First",
        "tags" : [ "Greeting Resource" ]
      }
    },
    "/schema/UseSchemaRefType" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UseSchemaRefType"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Second",
        "tags" : [ "Greeting Resource" ]
      }
    }
  },
  "info" : {
    "title" : "openapi-schema-problems API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}